% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBoost.R
\name{L1GBoost}
\alias{L1GBoost}
\title{Group Boosting}
\usage{
L1GBoost(X, y, iter, n_per, weight_p)
}
\arguments{
\item{X}{matrix of regressor variables}

\item{y}{depedent / repsonse variable}

\item{iter}{nubmer of iterations}

\item{n_per}{number of observations per unit (size of gropups)}

\item{weight_p}{weight function}
}
\value{
A list with different components comparable to \code{L2Boost} is returned.
}
\description{
This functions impelements an algorithm for group L2Boosting with a L1 loss function (median loss).
}

