% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost.R
\name{L2Boost}
\alias{L2Boost}
\title{Implementation of the L2Boosting algorithm}
\usage{
L2Boost(X, y, iter = 200, beta.start = rep(0, dim(X)[2]), post = FALSE)
}
\arguments{
\item{X}{matrix of regressors}

\item{y}{dependent variable}

\item{iter}{number of iterations}

\item{beta.start}{initial value for the algorithm to start with}

\item{post}{logical, if post Boosting algorithm should be applied (default \code{FALSE}).}
}
\value{
The functions returns an object of class \code{L2Boost} with the following components:
\itemize{
\item{BetaFinal}{the estimated final parameter vector.}
\item{BetaVector}{vector where each component gives the values of the estimated coefficient in each round.}
\item{Res}{matrix with the estimated function values in each round}
\item{Res_beta}{matrix with the estimated coefficient vectors from each round}
 \item{i}{Iteration when the process was terminated.}
 \item{S}{vector of the indices of the selected variables in each round}
  \item{sigma2}{estimation of the variance} 
  \item{stop_rule}{When the stopping rule, stopped the algorithm.}
  \item{iter}{maximal number of iterations of the algorithm}
  \item{y}{model response / dependent variable}
  \item{X}{model matrix}
}
}
\description{
This function implements a classical version of the L2Boosting algorithm.
}

