% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost.R
\name{MSECal}
\alias{MSECal}
\title{Function for calculation of different MSEs}
\usage{
MSECal(object, yref, beta.true = NULL)
}
\arguments{
\item{object}{object of class \code{L2Boost} or \code{L2BoostOGA}}

\item{yref}{reference variable with regard to which the MSE is calculated (e.g. the observed dependent variables or some oracle)}

\item{beta.true}{the ture beta coefficient if known}
}
\value{
The function returns a list with different measures of th
}
\description{
This function is mainly used for analyzing the performance of the L2Boost and orthogonal L2Boost algorithms.
}

