% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost.R
\name{predict.L2BoostOGA}
\alias{predict.L2BoostOGA}
\title{Predict fot L2BoostOGA objects}
\usage{
\method{predict}{L2BoostOGA}(object, newx = NULL, mstop = object$stop_rule,
  ...)
}
\arguments{
\item{object}{Object of class \code{L2BoostOGA}}

\item{newx}{Matrix of x-variables to be predicted}

\item{mstop}{Stopping number, i.e. for which iteration of the algorithm the prediction is to be conducted}

\item{...}{further arguments}
}
\value{
A vector with the predicted values.
}
\description{
Predicted values based on L2BoostOGA objects
}

